Add snippets
