+++ OpenRPA.Forms/Snippets/SimpleForm.xaml
@@ -1,0 +1,55 @@
+﻿<Activity mc:Ignorable="sap sap2010 sads" x:Class="SimpleForm" mva:VisualBasic.Settings="{x:Null}" sap2010:WorkflowViewState.IdRef="FormTest_1"
+ xmlns="http://schemas.microsoft.com/netfx/2009/xaml/activities"
+ xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+ xmlns:mva="clr-namespace:Microsoft.VisualBasic.Activities;assembly=System.Activities"
+ xmlns:ofa="clr-namespace:OpenRPA.Forms.Activities;assembly=OpenRPA.Forms"
+ xmlns:sads="http://schemas.microsoft.com/netfx/2010/xaml/activities/debugger"
+ xmlns:sap="http://schemas.microsoft.com/netfx/2009/xaml/activities/presentation"
+ xmlns:sap2010="http://schemas.microsoft.com/netfx/2010/xaml/activities/presentation"
+ xmlns:scg="clr-namespace:System.Collections.Generic;assembly=mscorlib"
+ xmlns:sco="clr-namespace:System.Collections.ObjectModel;assembly=mscorlib"
+ xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    <TextExpression.NamespacesForImplementation>
+        <sco:Collection x:TypeArguments="x:String">
+            <x:String>System</x:String>
+            <x:String>Microsoft.VisualBasic</x:String>
+            <x:String>System.Xml</x:String>
+            <x:String>OpenRPA</x:String>
+            <x:String>System.Data</x:String>
+            <x:String>System.Linq</x:String>
+            <x:String>OpenRPA.IE</x:String>
+            <x:String>OpenRPA.Image</x:String>
+            <x:String>OpenRPA.Windows</x:String>
+        </sco:Collection>
+    </TextExpression.NamespacesForImplementation>
+    <TextExpression.ReferencesForImplementation>
+        <sco:Collection x:TypeArguments="AssemblyReference">
+            <AssemblyReference>mscorlib</AssemblyReference>
+            <AssemblyReference>Microsoft.VisualBasic</AssemblyReference>
+            <AssemblyReference>System.Xml</AssemblyReference>
+            <AssemblyReference>OpenRPA</AssemblyReference>
+            <AssemblyReference>OpenRPA.Interfaces</AssemblyReference>
+            <AssemblyReference>System.Data</AssemblyReference>
+            <AssemblyReference>System.Core</AssemblyReference>
+            <AssemblyReference>OpenRPA.Office</AssemblyReference>
+            <AssemblyReference>OpenRPA.Windows</AssemblyReference>
+        </sco:Collection>
+    </TextExpression.ReferencesForImplementation>
+    <Sequence DisplayName="SimpleForm" sap2010:WorkflowViewState.IdRef="Sequence_1">
+        <ofa:InvokeForm DisplayName="Patient Data" Form="&lt;form&gt;&#xA;    &lt;title&gt;Create account&lt;/title&gt;&#xA;    &lt;heading&gt;Personal details&lt;/heading&gt;&#xA;    &lt;input type='string' name='FirstName' &#xA;           label='First name'&#xA;           tooltip='Enter your first name here.'&#xA;           icon='pencil'&gt;&#xA;        &lt;validate must='NotBeEmpty' message='First name is mandatory' /&gt;&#xA;    &lt;/input&gt;&#xA;    &lt;input type='string' name='LastName'&#xA;           label='Last name' icon='pencil'&#xA;           tooltip='Enter your last name here.'&gt;&#xA;        &lt;validate must='NotBeEmpty' message='Last name is mandatory' /&gt;&#xA;    &lt;/input&gt;&#xA;    &lt;input type='int' name='Age'&#xA;           label='Your age' icon='PencilOff' &#xA;           tooltip='Enter your age here.' &gt;&#xA;        &lt;validate must='NotBeEmpty' message='Age is mandatory' /&gt;&#xA;        &lt;validate must='BeLessThan' value='110' message='Age must be less than 110' /&gt;&#xA;        &lt;validate must='BeGreaterThan' value='18' message='Age must be greater than 18' /&gt;&#xA;    &lt;/input&gt;&#xA;&#xA;    &lt;action name='submit' content='SUBMIT' icon='check' validates='true' ClosesDialog='true' /&gt;&#xA;&lt;/form&gt;" sap2010:WorkflowViewState.IdRef="InvokeForm_1" />
+        <sads:DebugSymbol.Symbol>dw1DOlxFbXB0eS54YW1sAA==</sads:DebugSymbol.Symbol>
+    </Sequence>
+    <sap2010:WorkflowViewState.ViewStateManager>
+        <sap2010:ViewStateManager>
+            <sap2010:ViewStateData Id="InvokeForm_1" sap:VirtualizedContainerService.HintSize="200,74" />
+            <sap2010:ViewStateData Id="Sequence_1" sap:VirtualizedContainerService.HintSize="222,198">
+                <sap:WorkflowViewStateService.ViewState>
+                    <scg:Dictionary x:TypeArguments="x:String, x:Object">
+                        <x:Boolean x:Key="IsExpanded">True</x:Boolean>
+                    </scg:Dictionary>
+                </sap:WorkflowViewStateService.ViewState>
+            </sap2010:ViewStateData>
+            <sap2010:ViewStateData Id="FormTest_1" sap:VirtualizedContainerService.HintSize="262,278" />
+        </sap2010:ViewStateManager>
+    </sap2010:WorkflowViewState.ViewStateManager>
+</Activity>
