+++ OpenRPA.IE/Extensions.cs
@@ -9,16 +9,16 @@ using System.Windows.Forms;
 
 namespace OpenRPA.IE
 {
-    public static class Extensions
+    public static class IEExtensions
     {
         public static System.Windows.Media.Imaging.BitmapFrame GetImageSourceFromResource(string resourceName)
         {
-            string[] names = typeof(Extensions).Assembly.GetManifestResourceNames();
+            string[] names = typeof(IEExtensions).Assembly.GetManifestResourceNames();
             foreach (var name in names)
             {
                 if (name.EndsWith(resourceName))
                 {
-                    return System.Windows.Media.Imaging.BitmapFrame.Create(typeof(Extensions).Assembly.GetManifestResourceStream(name));
+                    return System.Windows.Media.Imaging.BitmapFrame.Create(typeof(IEExtensions).Assembly.GetManifestResourceStream(name));
                 }
             }
             return null;
