+++ OpenRPA.NM/OpenRPA.NM.csproj
@@ -26,6 +26,7 @@
     <None Remove="Activities\OpenURLDesigner.xaml" />
     <None Remove="Resources\designer\gethtmlelement.png" />
     <None Remove="Resources\toolbox\gethtmlelement.png" />
+    <None Remove="Snippets\GoogleSearch.xaml" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\toolbox\gethtmlelement.png" />
@@ -53,6 +54,10 @@
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Resources\designer\gethtmlelement.png" />
+    <Resource Include="Snippets\GoogleSearch.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Resource>
   </ItemGroup>
   <Target Name="PostBuild" AfterTargets="Pack" Condition="'$(Configuration)' == 'Debug'">
     <Exec Command="copy $(OutputPath)..\$(PackageId).$(PackageVersion).nupkg C:\code\OpenRPA\packages" />
