+++ OpenRPA.Office/Activities/CloseWorkbookDesigner.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Activities.Presentation.Model;
+﻿using OpenRPA.Interfaces;
+using System.Activities.Presentation.Model;
 using System.Windows;
 using System.Windows.Media;
