+++ OpenRPA.Office/Activities/ReadCellDesigner.xaml.cs
@@ -1,4 +1,5 @@
 ﻿//using rpaactivities.activityextension;
+using OpenRPA.Interfaces;
 using System;
 using System.Activities.Presentation.View;
 using System.ComponentModel;
