--- OpenRPA.Office/Extensions.cs
@@ -1,47 +1,0 @@
-﻿using FlaUI.Core.AutomationElements.Infrastructure;
-using OpenRPA.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Management;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace OpenRPA.Office
-{
-    public static class Extensions
-    {
-        public static string replaceEnvironmentVariable(this string filename)
-        {
-            var USERPROFILE = Environment.GetEnvironmentVariable("USERPROFILE");
-            var windir = Environment.GetEnvironmentVariable("windir");
-            var SystemRoot = Environment.GetEnvironmentVariable("SystemRoot");
-            var PUBLIC = Environment.GetEnvironmentVariable("PUBLIC");
-
-            if (!string.IsNullOrEmpty(USERPROFILE)) filename = filename.Replace(USERPROFILE, "%USERPROFILE%");
-            if (!string.IsNullOrEmpty(windir)) filename = filename.Replace(windir, "%windir%");
-            if (!string.IsNullOrEmpty(SystemRoot)) filename = filename.Replace(SystemRoot, "%SystemRoot%");
-            if (!string.IsNullOrEmpty(PUBLIC)) filename = filename.Replace(PUBLIC, "%PUBLIC%");
-
-            var ProgramData = Environment.GetEnvironmentVariable("ProgramData");
-            if (!string.IsNullOrEmpty(ProgramData)) filename = filename.Replace(ProgramData, "%ProgramData%");
-            var ProgramFilesx86 = Environment.GetEnvironmentVariable("ProgramFiles(x86)");
-            if (!string.IsNullOrEmpty(ProgramFilesx86)) filename = filename.Replace(ProgramFilesx86, "%ProgramFiles(x86)%");
-            var ProgramFiles = Environment.GetEnvironmentVariable("ProgramFiles");
-            if (!string.IsNullOrEmpty(ProgramFiles)) filename = filename.Replace(ProgramFiles, "%ProgramFiles%");
-            var LOCALAPPDATA = Environment.GetEnvironmentVariable("LOCALAPPDATA");
-            if (!string.IsNullOrEmpty(LOCALAPPDATA)) filename = filename.Replace(LOCALAPPDATA, "%LOCALAPPDATA%");
-            var APPDATA = Environment.GetEnvironmentVariable("APPDATA");
-            if (!string.IsNullOrEmpty(APPDATA)) filename = filename.Replace(APPDATA, "%APPDATA%");
-
-
-            //var = Environment.GetEnvironmentVariable("");
-            //if (!string.IsNullOrEmpty()) filename = filename.Replace(, "%%");
-
-            return filename;
-        }
-    }
-
-}
