+++ OpenRPA.Script/Activities/ExecuteAHKScriptDesigner.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Activities;
+﻿using OpenRPA.Interfaces;
+using System.Activities;
 using System.Activities.Expressions;
 using System.Activities.Presentation.Model;
 using System.Windows;
