+++ OpenRPA.Snippets/ExcelLoop.cs
@@ -1,32 +1,20 @@
-﻿using System;
-using System.Activities;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using OpenRPA.Interfaces;
-using OpenRPA.Interfaces.entity;
+﻿//using System;
+//using System.Activities;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Text;
+//using System.Threading.Tasks;
+//using System.Windows;
+//using OpenRPA.Interfaces;
+//using OpenRPA.Interfaces.entity;
 
-namespace OpenRPA.Snippets
-{
-    public class ExcelLoop : Interfaces.ISnippet
-    {
-        public string Name => "ExcelLoop";
-        public string Category => "Excel";
-        public string Xaml => Extensions.ResourceAsString("ExcelLoop.xaml");
-        public snippet Snippet { get; set; }
-        public Activity Create(DependencyObject target)
-        {
-            try
-            {
-                return System.Xaml.XamlServices.Load(new System.IO.StringReader(Xaml)) as Activity;
-            }
-            catch (Exception ex)
-            {
-                Log.Error(ex.ToString());
-                throw;
-            }
-        }
-    }
-}
+//namespace OpenRPA.Snippets
+//{
+//    public class ExcelLoop : ISnippet
+//    {
+//        public string Name => "ExcelLoop";
+//        public string Category => "Excel";
+//        public string Xaml => Extensions.ResourceAsString(typeof(ExcelLoop), "ExcelLoop.xaml");
+//        public snippet Snippet { get; set; }
+//    }
+//}
